Few refactors
- Use stdint types
- Rename for consistency
- Few leaks

+++ src/utils.c
@@ -1,6 +1,9 @@
 #include "utils.h"
 
-unsigned int random_in_range(unsigned int min, unsigned int max)
+#include <stdint.h>
+#include <stdlib.h>
+
+uint64_t random_in_range(uint64_t min, uint64_t max)
 {
     return rand() % (max + 1 - min) + min;
 }
